+++ app/src/main/java/com/silverknife/meizhi/mvp/ui/activity/PhotoViewActivity.kt
@@ -1,6 +1,7 @@
 package com.silverknife.meizhi.mvp.ui.activity
 
 import android.text.TextUtils
+import androidx.core.app.ActivityCompat
 import com.silvericekey.skutilslibrary.base.BaseActivity
 import com.silvericekey.skutilslibrary.utils.ImageLoderUtil
 import com.silverknife.meizhi.R
@@ -10,6 +11,10 @@ import kotlinx.android.synthetic.main.gank_list_item.*
 import kotlinx.android.synthetic.main.gank_list_item.content_title
 
 class PhotoViewActivity : BaseActivity<PhotoViewPresenter>() {
+    companion object {
+        val IMAGE = "image"
+    }
+
     override fun getLayoutID(): Int {
         return R.layout.activity_photo_view
     }
@@ -21,9 +26,10 @@ class PhotoViewActivity : BaseActivity<PhotoViewPresenter>() {
         } else {
             content_title.text = "图片"
         }
-        back.setOnClickListener { finish() }
+        back.setOnClickListener { ActivityCompat.finishAfterTransition(this) }
         var url = intent.getStringExtra("url")
         ImageLoderUtil.bindImg(this, url, ImageLoderUtil.Builder().build()).into(image)
+        addTransitionName(image, IMAGE)
     }
 
     override fun initPresenter(): PhotoViewPresenter {