+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - ```2019/09/29``` BaseFragment新增onVisibale判断是否从后台切换回前台，取消fragment下presenter的lateinit让onVisible判断标准为非初次加载，自动集成MultiDex，让Application继承BaseApplication即可，最小SDK支持变更为16，GlueTabLayout中修改Icon为非必传，demo界面修改，新增干货集中营API，待完善。
 - ```2019/09/30``` 修复福利界面图片无法加载的bug，修复多图片recyclerview空白处无法点击跳转的bug，新增图片预览功能，使用框架photoview，新增详情页跳转。
 - ```2019/10/09``` 添加navigationview侧滑菜单，主界面改为fragment的切换。
+- ```2019/10/14``` BaseActivity，BaseFragment 增加initOptionsCompat方法用于添加转场动画，重写startActivity,判断转场动画参数optionsCompat:ActivityOptionsCompat是否为null，如果不为null则增加，BaseActivity增加addTransitionName方法用于设置被跳转activity的转场view。